figure out how relative addresses work.
Basically the important part is that it is relative to the CIE (really FDE)
that is currently being processed, in addition to relative to a
few constants that somewhat make sense: the load address and the
offset of .eh_frame within the ELF image.

The fun part is that I have never seen a document that describes this
calculation.  I derived the above by comparing against 'readelf
--debug-dump=frames's output.