implement some basic DwarfVM instructions, + tests.
This adds a braindead simple binary as a test program, and adds
some test code that parses out the exception handling information
from every FDE in it.  It's a simple test to make sure we are
parsing, and to some extent, interpreting the instructions
correctly.

The binary is very tiny and should not change much, if at all.