Don't export 'Register', ignore some instructions.
We'll use "Register" as an actual exported thing in a second, but
the exported set of values is more limited than what this
'Register' is, so rename to 'register' to clearly demarcate the
two sets and make the interface have a smaller valid set of
registers.

Since our hacky setup only cares about the canonical frame
address, we can ignore a bunch of instructions (hopefully).
Properly ignore some of them here.