start of the code that will be the modules interface.
Hope to just expose 'RetAddrAddr', in the end.  Maybe that should
be renamed 'CFA' or something.

Includes a basic test to make sure we interpret obvious
instructions correctly.  The given test actually just executes the
CIE; the FDE program for that entry is empty (not to mention we
request the entry address, anyway).